[FIX] warnings in swap_ks_manager
build it with -Werror

it depends on fix to swap_kprobe

+++ ks_manager/ks_manager.c
@@ -71,7 +71,7 @@ static struct probe *find_probe(unsigned long addr)
 
 	/* check if such probe does exist */
 	swap_hlist_for_each_entry(p, node, &list_probes, hlist)
-		if (p->p.jp.kp.addr == addr)
+		if ((unsigned long)p->p.jp.kp.addr == addr)
 			return p;
 
 	return NULL;
